V1.06 - New UIN Gen & Table Population Fixes
User Management Sprint - Unique Identification Number is now done by taking the UIN of the last entry on the list and adding one, the previous method could lead to erros when deleting, creating 2 users with the same UIN.  Table on AddUser page now populates properly, as previously it would only populate with one entry.

JForms made for each of the Roles.